requirements.txt
streamlit
pandas
langchain
langchain-google-genai
langchain-experimental
plotly
seaborn
matplotlib
python-docx
numpy
tabulate
kaleido
python-docx

